SC-5213 extract helpdesk from administration context
