+++ controllers/administration.js
@@ -92,52 +92,6 @@ const getTableActions = (
 	];
 };
 
-const getTableActionsSend = (item, path, state, res) => {
-	const actions = [];
-	if (state === 'submitted' || state === 'closed') {
-		actions.push(
-			{
-				link: path + item._id,
-				class: 'btn-edit',
-				icon: 'edit',
-				title: res.$t('administration.controller.link.editEntry'),
-			},
-			{
-				class: 'disabled',
-				icon: 'archive',
-			},
-			{
-				class: 'disabled',
-				icon: 'paper-plane',
-			},
-		);
-	} else {
-		actions.push(
-			{
-				link: path + item._id,
-				class: 'btn-edit',
-				icon: 'edit',
-				title: res.$t('administration.controller.link.editEntry'),
-			},
-			{
-				link: path + item._id,
-				class: 'btn-disable',
-				icon: 'archive',
-				method: 'delete',
-				title: res.$t('administration.controller.link.completeEntry'),
-			},
-			{
-				link: path + item._id,
-				class: 'btn',
-				icon: 'paper-plane',
-				method: 'post',
-				title: res.$t('administration.controller.link.sendEntryToDevelopmentTeam'),
-			},
-		);
-	}
-	return actions;
-};
-
 /**
  * maps the event props from the server to fit the ui components, e.g. date and time
  * @param data {object} - the plain data object
@@ -458,56 +412,6 @@ const getUserCreateHandler = (internalReturn) => function userCreate(req, res, n
 				}),
 			};
 			return redirectHelper.safeBackRedirect(req, res);
-		});
-};
-
-/**
- * send out problem to the sc helpdesk
- * @param service currently only used for helpdesk
- * @returns {Function}
- */
-const getSendHelper = (service) => function send(req, res, next) {
-	api(req)
-		.get(`/${service}/${req.params.id}`)
-		.then((data) => {
-			const user = res.locals.currentUser;
-
-			api(req)
-				.post('/helpdesk', {
-					json: {
-						type: 'contactHPI',
-						subject: data.subject,
-						role: '',
-						desire: '',
-						benefit: '',
-						acceptanceCriteria: '',
-						currentState: data.currentState,
-						targetState: data.targetState,
-						notes: data.notes,
-						schoolName: res.locals.currentSchoolData.name,
-						userId: user._id,
-						email: user.email ? user.email : '',
-						schoolId: res.locals.currentSchoolData._id,
-						cloud: res.locals.theme.title,
-						browserName: '',
-						browserVersion: '',
-						os: '',
-						device: '',
-						deviceUserAgent: '',
-					},
-				})
-				.then(() => {
-					api(req).patch(`/${service}/${req.params.id}`, {
-						json: {
-							state: 'submitted',
-							order: 1,
-						},
-					});
-				})
-				.catch((err) => {
-					res.status(err.statusCode || 500).send(err);
-				});
-			redirectHelper.safeBackRedirect(req, res);
 		});
 };
 
@@ -2359,127 +2263,6 @@ router.get(
 					pagination,
 					limit: true,
 					filterSettings: JSON.stringify(classFilterSettings({ years, currentYear }, res)),
-				});
-			});
-	},
-);
-
-/**
- * Set state to closed of helpdesk problem
- * @param service usually helpdesk, to disable instead of delete entry
- * @returns {Function}
- */
-const getDisableHandler = (service) => function diasableHandler(req, res, next) {
-	api(req)
-		.patch(`/${service}/${req.params.id}`, {
-			json: {
-				state: 'closed',
-				order: 2,
-			},
-		})
-		.then(() => {
-			redirectHelper.safeBackRedirect(req, res);
-		});
-};
-
-/**
- * Truncates string to 25 chars
- * @param string given string to truncate
- * @returns {string}
- */
-const truncate = (string) => {
-	if ((string || {}).length > 25) {
-		return `${string.substring(0, 25)}...`;
-	}
-	return string;
-};
-
-/*
-    HELPDESK
-*/
-
-router.patch(
-	'/helpdesk/:id',
-	permissionsHelper.permissionsChecker('HELPDESK_VIEW'),
-	getUpdateHandler('helpdesk'),
-);
-router.get(
-	'/helpdesk/:id',
-	permissionsHelper.permissionsChecker('HELPDESK_VIEW'),
-	getDetailHandler('helpdesk'),
-);
-router.delete(
-	'/helpdesk/:id',
-	permissionsHelper.permissionsChecker('HELPDESK_VIEW'),
-	getDisableHandler('helpdesk'),
-);
-router.post(
-	'/helpdesk/:id',
-	permissionsHelper.permissionsChecker('HELPDESK_VIEW'),
-	getSendHelper('helpdesk'),
-);
-router.all(
-	'/helpdesk',
-	permissionsHelper.permissionsChecker('HELPDESK_VIEW'),
-	(req, res, next) => {
-		const itemsPerPage = req.query.limit || 10;
-		const currentPage = parseInt(req.query.p, 10) || 1;
-		const title = returnAdminPrefix(res.locals.currentUser.roles, res);
-
-		api(req)
-			.get('/helpdesk', {
-				qs: {
-					$limit: itemsPerPage,
-					$skip: itemsPerPage * (currentPage - 1),
-					$sort: req.query.sort ? req.query.sort : { order: 1 },
-					schoolId: res.locals.currentSchool,
-				},
-			})
-			.then((data) => {
-				const head = [
-					res.$t('global.label.title'),
-					res.$t('administration.controller.headline.itsOn'),
-					res.$t('administration.controller.headline.targetState'),
-					res.$t('administration.controller.headline.status'),
-					res.$t('administration.controller.headline.creationDate'),
-					res.$t('administration.controller.headline.remarks'),
-					'',
-				];
-
-				const body = data.data.map((item) => [
-					truncate(item.subject || ''),
-					truncate(item.currentState || ''),
-					truncate(item.targetState || ''),
-					res.$t(`administration.controller.text.${item.state}`),
-					moment(item.createdAt).format('DD.MM.YYYY'),
-					truncate(item.notes || ''),
-					getTableActionsSend(item, '/administration/helpdesk/', item.state, res),
-				]);
-
-				let sortQuery = '';
-				if (req.query.sort) {
-					sortQuery = `&sort=${req.query.sort}`;
-				}
-
-				let limitQuery = '';
-				if (req.query.limit) {
-					limitQuery = `&limit=${req.query.limit}`;
-				}
-
-				const pagination = {
-					currentPage,
-					numPages: Math.ceil(data.total / itemsPerPage),
-					baseUrl: `/administration/helpdesk/?p={{page}}${sortQuery}${limitQuery}`,
-				};
-
-				res.render('administration/helpdesk', {
-					title: res.$t('administration.controller.headline.helpdesk', {
-						title,
-					}),
-					head,
-					body,
-					pagination,
-					limit: true,
 				});
 			});
 	},
