+++ views/helpdesk/forms/form-problem.hbs
@@ -1,0 +1,26 @@
+<input name="schoolId" type="hidden" data-force-value="true" value="{{currentSchool}}" />
+
+<div class="form-group">
+    <label>{{$t "administration.form_problem.label.createdOn" }} </label>
+    <textarea class="form-control" name="createdAt" type="text" rows="1" disabled></textarea>
+</div>
+
+<div class="form-group">
+    <label>{{$t "global.label.title" }}</label>
+    <input class="form-control" name="subject" type="text" disabled />
+</div>
+
+<div class="form-group">
+    <label>{{$t "administration.form_problem.label.itsOn" }}</label>
+    <textarea class="form-control" name="currentState" type="text" cols="50" rows="6" disabled></textarea>
+</div>
+
+<div class="form-group">
+    <label>{{$t "administration.form_problem.label.targetState" }}</label>
+    <textarea class="form-control" name="targetState" type="text" cols="50" rows="6" disabled></textarea>
+</div>
+
+<div class="form-group">
+    <label>{{$t "administration.form_problem.label.remarks" }}</label>
+    <textarea class="form-control" name="notes" type="text" cols="50" rows="6"></textarea>
+</div>
