+++ views/helpdesk/helpdesk.hbs
@@ -1,0 +1,15 @@
+{{#extend "lib/loggedin"}}
+    {{#content "scripts" mode="append"}}
+        <script src="/scripts/administration.js" type="text/javascript" nonce="{{nonceValue}}" defer></script>
+        <script src="/scripts/tables.js" type="text/javascript" nonce="{{nonceValue}}" defer></script>
+        <script src="/scripts/filter/custom-element-polyfill.js" nonce="{{nonceValue}}" defer></script>
+        <script src="/scripts/filter/feathers-filter.js" nonce="{{nonceValue}}" defer></script>
+        <script src="/scripts/base64files.js" type="text/javascript" nonce="{{nonceValue}}" defer></script>
+    {{/content}}
+    {{#content "styles" mode="append"}}
+        <link rel="stylesheet" href="/styles/lib/datetimepicker/jquery.datetimepicker.min.css"/>
+    {{/content}}
+    {{#content "page"}}
+        {{> "administration/partials/data_protection_hint"}}
+    {{/content}}
+{{/extend}}
