+++ views/helpdesk/overview.hbs
@@ -1,0 +1,34 @@
+{{#extend "helpdesk/helpdesk"}}
+    {{#content "page" mode="prepend"}}
+        <div class="route-administration">
+            <section class="section-school section-default">
+                <div class="container-fluid">
+                    <div class="row">
+                        <div class="col-sm-12 no-padding">
+                            {{> "lib/components/table" head=head body=body}}
+                        </div>
+                    </div>
+                    <div class="row">
+                        {{> "lib/components/pagination" settings=pagination }}
+                    </div>
+                </div>
+
+                {{#embed "lib/components/modal-form" method="patch" class="edit-modal"}}
+                    {{#content "fields"}}
+                        {{> "helpdesk/forms/form-problem"}}
+                    {{/content}}
+                {{/embed}}
+
+                <p>
+                    {{{$t "administration.longText.text.provideStudentsAndTheirParents" (dict "title" theme.title "shortTitle" theme.short_title)}}}
+                    <p>
+                        <a href="/help?activeForm=team#contact-form" class="btn btn-primary">
+                        {{$t "administration.global.button.contactSupport" }}
+                        </a>
+                    </p>
+                    {{{$t "administration.global.helpdesk.label.tips" }}}
+                </p>
+            </section>
+        </div>
+    {{/content}}
+    {{/extend}}
